+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.2.1] - 2026-09-19
+
+### Added
+- More content for 'Class Skills'.
+
+### Internal
+- `app.js` reorganized and given more guardrails.
+
+---
+
 ## [0.2.0] - 2026-09-18
 
 ### Added